refactor(scraper): Rewrite `scrapeStudentSchedule` -> `populateStudentWithLessons`
